Use form to populate object
